+++ command/craft/craft.go
@@ -108,8 +108,12 @@ func NewCommand(home string, docker client.CommonAPIClient, output terminal.Outp
 
 			// start the container if its not running
 			if containers[0].State != "running" {
-				if err := docker.ContainerStart(cmd.Context(), containers[0].ID, types.ContainerStartOptions{}); err != nil {
-					return err
+				for _, command := range cmd.Root().Commands() {
+					if command.Use == "start" {
+						if err := command.RunE(cmd, []string{}); err != nil {
+							return err
+						}
+					}
 				}
 			}