+++ command/database/import.go
@@ -116,7 +116,7 @@ func importCommand(home string, docker client.CommonAPIClient, nitrod protob.Nit
 			}
 
 			// get a list of all the databases
-			containers, err := docker.ContainerList(cmd.Context(), types.ContainerListOptions{Filters: filter})
+			containers, err := docker.ContainerList(cmd.Context(), types.ContainerListOptions{Filters: filter, All: true})
 			if err != nil {
 				return err
 			}
@@ -130,8 +130,12 @@ func importCommand(home string, docker client.CommonAPIClient, nitrod protob.Nit
 			var options []string
 			for _, c := range containers {
 				if c.State != "running" {
-					if err := docker.ContainerStart(cmd.Context(), c.ID, types.ContainerStartOptions{}); err != nil {
-						return err
+					for _, command := range cmd.Root().Commands() {
+						if command.Use == "start" {
+							if err := command.RunE(cmd, []string{}); err != nil {
+								return err
+							}
+						}
 					}
 				}